merged from upstream

+++ tools/setup_Debian.sh
@@ -20,15 +20,19 @@ function add_key {
 }
 
 function setup_disk {
-    sudo swapoff -a
-    sudo dd if=/dev/zero of=/swapfile bs=1M count=4096
-    sudo chmod 0600 /swapfile
-    sudo mkswap /swapfile
-    sudo /sbin/swapon /swapfile
+    if [ ! -f /swapfile ]; then
+        sudo swapoff -a
+        sudo dd if=/dev/zero of=/swapfile bs=1M count=4096
+        sudo chmod 0600 /swapfile
+        sudo mkswap /swapfile
+        sudo /sbin/swapon /swapfile
+    fi
 
-    sudo dd if=/dev/zero of=/docker bs=1M count=10240
-    losetup -f /docker
-    DEV=$(losetup -a | awk -F: '/\/docker/ {print $1}')
+    if [ ! -f /docker ]; then
+        sudo dd if=/dev/zero of=/docker bs=1M count=10240
+        losetup -f /docker
+        DEV=$(losetup -a | awk -F: '/\/docker/ {print $1}')
+    fi
 
     # Format Disks and setup Docker to use BTRFS
     sudo parted ${DEV} -s -- mklabel msdos
@@ -51,7 +55,7 @@ export PATH="/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
 echo 'deb http://apt.dockerproject.org/repo ubuntu-trusty main' | sudo tee /etc/apt/sources.list.d/docker.list
 add_key
 sudo apt-get update
-sudo apt-get install -y --no-install-recommends docker-engine btrfs-tools
+sudo apt-get -y install --no-install-recommends docker-engine btrfs-tools
 
 sudo service docker stop
 setup_disk
